# apps/whatsapp-api/docs/api-documentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Next Mavens Fidscript WhatsApp API",
    "description": "Production-ready WhatsApp Business API with anti-ban protection, rate limiting, and latest WhatsApp features. Multi-tenant SaaS platform for sending and receiving WhatsApp messages programmatically.",
    "version": "1.0.0",
    "contact": {
      "name": "Next Mavens Support",
      "email": "support@nextmavens.com"
    },
    "license": {
      "name": "Commercial",
      "url": "https://nextmavens.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development"
    }
  ],
...
  "paths": {
...
    "/campaign/create": {
      "post": {
        "tags": ["Campaign"],
        "summary": "Create a new campaign (metadata only)",
        "security": [{"ApiKeyAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "instanceName"],
                "properties": {
                  "name": {"type": "string"},
                  "instanceName": {"type": "string"},
                  "description": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Campaign created successfully"}
        }
      }
    },
    "/campaign/schedule": {
      "post": {
        "tags": ["Campaign"],
        "summary": "Schedule a campaign for future sending with anti-ban protection",
        "security": [{"ApiKeyAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "instanceName", "message", "recipients"],
                "properties": {
                  "name": {"type": "string"},
                  "instanceName": {"type": "string"},
                  "description": {"type": "string"},
                  "message": {
                    "type": "object",
                    "properties": {
                      "type": {"type": "string", "enum": ["text", "media", "template", "interactive"]},
                      "text": {"type": "string"},
                      "mediaUrl": {"type": "string"},
                      "mediaType": {"type": "string"},
                      "mediaCaption": {"type": "string"},
                      "templateName": {"type": "string"},
                      "templateLanguage": {"type": "string"},
                      "templateComponents": {"type": "array"},
                      "interactiveButtons": {"type": "array"}
                    }
                  },
                  "recipients": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "phone": {"type": "string"},
                        "variables": {"type": "object"}
                      }
                    }
                  },
                  "scheduledAt": {"type": "string", "format": "date-time"},
                  "maxPerMinute": {"type": "integer", "minimum": 1, "maximum": 100}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Campaign scheduled successfully"}
        }
      }
    },
    "/campaign/send": {
      "post": {
        "tags": ["Campaign"],
        "summary": "Send a campaign immediately with anti-ban protection",
        "security": [{"ApiKeyAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "instanceName", "message", "recipients"],
                "properties": {
                  "name": {"type": "string"},
                  "instanceName": {"type": "string"},
                  "description": {"type": "string"},
                  "message": {"$ref": "#/components/schemas/CampaignMessage"},
                  "recipients": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "phone": {"type": "string"},
                        "variables": {"type": "object"}
                      }
                    }
                  },
                  "maxPerMinute": {"type": "integer", "minimum": 1, "maximum": 100}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "campaignId": {"type": "string"},
                    "status": {"type": "string"},
                    "sent": {"type": "integer"},
                    "failed": {"type": "integer"},
                    "pending": {"type": "integer"},
                    "blocked": {"type": "integer"},
                    "rateLimited": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/campaign/status/{campaignId}": {
      "get": {
        "tags": ["Campaign"],
        "summary": "Get campaign status",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "campaignId", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "Campaign status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "campaignId": {"type": "string"},
                    "status": {"type": "string", "enum": ["pending", "scheduled", "running", "completed", "cancelled", "paused"]},
                    "totalRecipients": {"type": "integer"},
                    "sent": {"type": "integer"},
                    "failed": {"type": "integer"},
                    "pending": {"type": "integer"},
                    "startTime": {"type": "string"},
                    "endTime": {"type": "string"},
                    "nextSendTime": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/campaign/list": {
      "get": {
        "tags": ["Campaign"],
        "summary": "List all campaigns",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "instanceName", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Campaigns retrieved"}
        }
      }
    },
    "/campaign/pause/{campaignId}": {
      "post": {
        "tags": ["Campaign"],
        "summary": "Pause a running campaign",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "campaignId", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Campaign paused"}
        }
      }
    },
    "/campaign/resume/{campaignId}": {
      "post": {
        "tags": ["Campaign"],
        "summary": "Resume a paused campaign",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "campaignId", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Campaign resumed"}
        }
      }
    },
    "/campaign/cancel/{campaignId}": {
      "post": {
        "tags": ["Campaign"],
        "summary": "Cancel a campaign",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "campaignId", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Campaign cancelled"}
        }
      }
    },
    "/campaign/delete/{campaignId}": {
      "delete": {
        "tags": ["Campaign"],
        "summary": "Delete a campaign",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "campaignId", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "204": {"description": "Campaign deleted"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "API key authentication"
      }
    },
    "schemas": {
      "CampaignMessage": {
        "type": "object",
        "required": ["type"],
        "properties": {
          "type": {"type": "string", "enum": ["text", "media", "template", "interactive"]},
          "text": {"type": "string"},
          "mediaUrl": {"type": "string"},
          "mediaType": {"type": "string", "enum": ["image", "video", "document"]},
          "mediaCaption": {"type": "string"},
          "templateName": {"type": "string"},
          "templateLanguage": {"type": "string"},
          "templateComponents": {"type": "array"},
          "interactiveButtons": {"type": "array"}
        }
      }
    }
  }
}